passing mooneye's tima tests

+++ gameBoy/GPU.h
@@ -17,7 +17,7 @@ public:
 	uint8_t oam[0xA0] = {};//object attribute memory
 	uint8_t mode = 0;
 	uint8_t modeClock = 0;
-	int16_t cyclesPerScanline= 456;
+	int cyclesPerScanline= 456;
 	uint8_t vRamBank = 0;
 	void reset();
 	void connectToBus(BUS* bus);
